// src/app/models/users.ts
export interface Users {
    _id?: string;
    userCategory: string;
    userName: string;
    userSubcategory: string;
    file: File;
    userDescription?: string;
    department: string;
    numDep? :string
    region?: string;
    email: string;
    phoneNumber?: string;
    website?: string;
    password?: string;
    facebookLink?: string;
    twitterLink?: string;
    instagramLink?:string,
    youtubeLink?:string,
    spotifyLink?:string,
    deezerLink?:string
}